+++ processFile.h
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-std::priority_queue<HuffmanTreeNode>* parseFile(char *fileName, uint32_t* fileSize);
+std::priority_queue<HuffmanTreeNode>* parseFile(char *fileName, uint64_t* fileSize);
 std::vector<uint8_t>* encodeFile(char *fileName, std::map<uint8_t, std::vector<bool>> symbols);
 void compressFile(char *fileName);
 void decompressFile(char *fileName);
